requirements.txt
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
torchsummary
timm==0.9.2
connected-components-3d

numpy==1.26.4
matplotlib
nibabel
pandas
scipy
scikit-image
tqdm
pillow==10.2.0
six